backend/requirements.txt
# Requirements for MediTech_New Backend
# Generated: March 1, 2025

# FastAPI
# ใช้เวอร์ชันที่เก่ากว่าเพื่อหลีกเลี่ยงปัญหาการติดตั้ง
fastapi==0.95.2
uvicorn==0.22.0
# pydantic==2.4.2  # ต้องการ Rust compiler
pydantic==1.10.8  # เวอร์ชันเก่าที่ไม่ต้องการ Rust
# pydantic-settings==2.0.3  # ต้องการ pydantic 2.x
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.0.0

# Database
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
# psycopg2-binary==2.9.9  # Commented out due to build issues

# Testing
pytest==7.4.3
httpx==0.24.1
pytest-cov==4.1.0

# Utilities
tenacity==8.2.3
celery==5.3.4
redis==5.0.1
# Pillow==10.1.0  # Commented out due to build issues
python-dateutil==2.8.2

# AI/ML
# torch==2.1.1  # Commented out due to compatibility issues
# transformers==4.35.2  # Commented out due to dependency on torch
# scikit-learn==1.3.2  # Commented out due to build issues (requires Visual C++)
# pandas==2.1.3  # Commented out due to build dependencies
# numpy==1.26.2  # Commented out due to build dependencies

# Development
# black==23.11.0  # Commented out due to build issues
# isort==5.12.0  # Commented out due to build issues
# flake8==6.1.0  # Commented out due to build issues
# mypy==1.7.0  # Commented out due to build issues 